+++ src/dumped_var.rs
@@ -3,12 +3,3 @@ pub struct DumpedVar {
     pub value: usize,
     pub identifier: String,
 }
-
-impl DumpedVar {
-    pub fn new() -> Self {
-        DumpedVar {
-            value: 0,
-            identifier: "".to_string(),
-        }
-    }
-}
